refactor: RecordManager 객체 등을 삭제한다
변경 내용 :
RecordManager 인터페이스와 그 구현 객체를 삭제한다.
RecordManagerOfGameResultOfCar 객체의 기능을 GameRecorder 내부로 옮긴다.
