+++ src/main/java/racingcar/controller/GameController.java
@@ -4,7 +4,6 @@ import racingcar.domain.carfactory.CarFactory;
 import racingcar.domain.cars.Cars;
 import racingcar.domain.numbergenerator.RandomSingleDigitGenerator;
 import racingcar.domain.record.GameRecorder;
-import racingcar.domain.record.RecordManagerOfGameResultOfCar;
 import racingcar.domain.system.GameSystem;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
@@ -38,6 +37,6 @@ public class GameController {
     }
 
     private GameSystem createGameSystem(int gameRound) {
-        return new GameSystem(gameRound, new GameRecorder(new ArrayList<>(), new RecordManagerOfGameResultOfCar()));
+        return new GameSystem(gameRound, new GameRecorder(new ArrayList<>()));
     }
 }
