--- src/main/java/racingcar/domain/record/RecordManager.java
@@ -1,9 +1,0 @@
-package racingcar.domain.record;
-
-import racingcar.domain.cars.Cars;
-
-import java.util.List;
-
-public interface RecordManager<T> {
-    List<T> makeResults(int gameRound, Cars cars);
-}
