+++ src/test/java/racingcar/domain/system/GameSystemTest.java
@@ -8,7 +8,6 @@ import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.domain.car.Car;
 import racingcar.domain.cars.Cars;
 import racingcar.domain.record.GameRecorder;
-import racingcar.domain.record.RecordManagerOfGameResultOfCar;
 import racingcar.domain.result.GameResultOfCar;
 
 import java.util.ArrayList;
@@ -30,14 +29,14 @@ class GameSystemTest {
     @ValueSource(ints = {1, 5, 100})
     @DisplayName("생성자에 자연수인 게임 시도 횟수를 입력하였을 때, 오류가 발생하지 않는지 확인")
     void create_test(int gameRound) {
-        assertDoesNotThrow(() -> new GameSystem(gameRound, new GameRecorder(new ArrayList<>(), new RecordManagerOfGameResultOfCar())));
+        assertDoesNotThrow(() -> new GameSystem(gameRound, new GameRecorder(new ArrayList<>())));
     }
 
     @ParameterizedTest
     @ValueSource(ints = {-2, -1, 0})
     @DisplayName("생성자에 자연수가 아닌 게임 시도 횟수를 입력하였을 때, 오류를 발생시키는지 확인")
     void validate_error_test(int gameRound) {
-        assertThatThrownBy(() -> new GameSystem(gameRound, new GameRecorder(new ArrayList<>(), new RecordManagerOfGameResultOfCar())))
+        assertThatThrownBy(() -> new GameSystem(gameRound, new GameRecorder(new ArrayList<>())))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -46,7 +45,7 @@ class GameSystemTest {
     @MethodSource("getCarsOnSamePositionAndGameRound")
     @DisplayName("GameSystem 객체가 서로 다른 gameRound를 가질 때, executeRace 실행 시 finalRound의 값과 동일한 횟수만큼 position의 변화가 일어나는지 확인")
     void executeRace_test(Cars cars, int gameRound, int expectedPosition) {
-        GameSystem gameSystem = new GameSystem(gameRound, new GameRecorder(new ArrayList<>(), new RecordManagerOfGameResultOfCar()));
+        GameSystem gameSystem = new GameSystem(gameRound, new GameRecorder(new ArrayList<>()));
 
         gameSystem.executeRace(cars, () -> ENOUGH_POWER_TO_MOVE);
 
@@ -65,7 +64,7 @@ class GameSystemTest {
     @MethodSource("getCarsAndWinners")
     @DisplayName("getWinnersGameResult 메소드가 우승 자동차들의 게임결과 객체를 반환하는지 확인")
     void getWinnersGameResult_test(Cars cars, List<String> expectedWinnerNames) {
-        GameSystem gameSystem = new GameSystem(DEFAULT_GAME_ROUND, new GameRecorder(new ArrayList<>(), new RecordManagerOfGameResultOfCar()));
+        GameSystem gameSystem = new GameSystem(DEFAULT_GAME_ROUND, new GameRecorder(new ArrayList<>()));
         gameSystem.executeRace(cars, () -> ENOUGH_POWER_TO_MOVE);
 
         List<GameResultOfCar> winnersGameResult = gameSystem.getWinnersGameResult();
